refactor: remove unused classes and reorganize package structure
