+++ src/main/kotlin/com/github/frozensync/command/cli/AbstractCommandCategory.kt
@@ -1,4 +1,4 @@
-package com.github.frozensync.discord.cli
+package com.github.frozensync.command.cli
 
 import discord4j.core.event.domain.message.MessageCreateEvent
