--- src/main/kotlin/com/github/frozensync/discord/command/Command.kt
@@ -1,5 +1,0 @@
-package com.github.frozensync.discord.command
-
-import discord4j.core.event.domain.message.MessageCreateEvent
-
-typealias Command = suspend (MessageCreateEvent) -> Unit
