--- src/main/kotlin/com/github/frozensync/discord/command/CommandSet.kt
@@ -1,5 +1,0 @@
-package com.github.frozensync.discord.command
-
-interface CommandSet {
-    val commands: Map<String, Command>
-}
